backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
numpy==2.2.3
python-dotenv==1.0.1
httpx>=0.27.0
# Groq Cloud (OpenAI-compatible)
groq==0.13.0
google-genai==1.29.0
# Supabase (Postgres)
supabase>=2.0.0
google-auth>=2.0.0
